+++ src/Client/ClientCore/Network/Pipeline/Inbound/ClientAllowedEventsInboundPipelineStage.cs
@@ -48,7 +48,8 @@ public class ClientAllowedEventsInboundPipelineStage : IInboundPipelineStage
         EventId.Core_Chat_System,
         EventId.Core_Chat_Generic,
         EventId.Core_Block_ModifyNotice,
-        EventId.Core_Block_RemoveNotice
+        EventId.Core_Block_RemoveNotice,
+        EventId.Core_Time_Clock
     };
 
     private readonly ILogger<ClientAllowedEventsInboundPipelineStage> logger;