+++ src/Server/ServerNetwork/Systems/Network/ServerOutboundEventSet.cs
@@ -41,6 +41,7 @@ public class ServerOutboundEventSet : IOutboundEventSet
         EventId.Core_Chat_System,
         EventId.Core_Chat_Generic,
         EventId.Core_Block_ModifyNotice,
-        EventId.Core_Block_RemoveNotice
+        EventId.Core_Block_RemoveNotice,
+        EventId.Core_Time_Clock
     };
 }